move definition to test-configs
